% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfIntersections.R
\name{any_intersect}
\alias{any_intersect}
\title{Check if a line has intersecting segments}
\usage{
any_intersect(line)
}
\arguments{
\item{line}{matrix of coordinates corresponding to the line of interest}
}
\value{
list where \code{list$any} is a boolean indicating if there are any intersections
and \code{list$val} is an index corresponding to the first intersection found
}
\description{
Determine if there are any intersecting line segments in a matrix of coordinates representing a line
}
\examples{
check_results <- any_intersect(currSecEx)
check_results$any #true/false
check_results$val #indices of first intersection found
}
