% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genContours.R
\name{censor}
\alias{censor}
\title{Truncate simulated line lengths based on a list of bounds}
\usage{
censor(x, bounds)
}
\arguments{
\item{x}{a vector of generated line lengths}

\item{bounds}{a vector which gives the lengths of the end and breakpoints
for each x value}
}
\value{
vector of new line lengths
}
\description{
Take in unbounded line lengths (x-values) and truncate them based on provided
bounds to create bounded lengths (y-values)
}
