% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfIntersections.R
\name{check_intersect}
\alias{check_intersect}
\title{Check if line segments intersect}
\usage{
check_intersect(a, b, c, d, seq = FALSE)
}
\arguments{
\item{a}{first coordinate of first line segment}

\item{b}{second coordinate of first line segment}

\item{c}{first coordinate of second line segment}

\item{d}{second coordinate of second line segment}

\item{seq}{indicator for whether the two line segments are intersecting}
}
\value{
boolean indicating if there is an intersection
}
\description{
Find if two line segments intersect
}
\examples{
 check_intersect(c(0, 0), c(1, 1), c(2, 2), c(3, 3))
 check_intersect(c(0, 0), c(1, 1), c(0.5, 0.5), c(2, 2))
}
