% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{clim_9_2008}
\alias{clim_9_2008}
\title{Climatology forecast for 2008.}
\format{array}
\usage{
data(clim_9_2008)
}
\description{
Proportion of times in the preceding ten years that sea ice concentration of
at least 15\%  was observed in each  grid box. Array of dimension year by
longitude by latitude. Computed from NASA Boostrap sea ice concentration
product.
}
\references{
Comiso, J., 2017: Bootstrap sea ice concentrations
            from Nimbus-7 SMMR and DMSP SSM/I-SSMIS. version 3.
            Boulder, Colorado USA: NASA National Snow and Ice Data Center
            Distributed Active Archive Center.
            doi: \url{https://doi.org/10.5067/7Q8HCCWS4I0R}
}
\keyword{datasets}
