% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF.R
\name{cond_prob}
\alias{cond_prob}
\title{Compute conditional probability of observed event}
\usage{
cond_prob(obs, mod)
}
\arguments{
\item{obs}{matrix with binary observations indicating if sea ice
concentration of at least 15$\%$ was observed. Dimension is lon x lat.}

\item{mod}{matrix with estimated sea ice probability from a model. Dimension
is lon x lat.}
}
\description{
Computes conditional probability of the observed event under some probability
model
}
