% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitContours.R
\name{contig_zero}
\alias{contig_zero}
\title{Find indices of sequences of contiguous zeros}
\usage{
contig_zero(y)
}
\arguments{
\item{y}{vector to consider}
}
\value{
Matrix of three columns where each row gives the first and last index
        of a sequence of numbers. The third column is a boolean. If
        \code{TRUE}, the indices are for a sequence zeros. If \code{FALSE},
        the indices are for a sequence non-zero values.
}
\description{
Identify the indices of sequences of repeated zero values and indices
of sequences of non-zero values
}
