% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importExport.R
\name{conv_to_grid}
\alias{conv_to_grid}
\title{Convert \code{SpatialPolygons} object to a grid}
\usage{
conv_to_grid(x, my_land_mat = land_mat)
}
\arguments{
\item{x}{\code{SpatialPolygons} object}

\item{my_land_mat}{binary matrix specifying land locations}
}
\description{
Convert \code{SpatialPolygons} object to a binary grid. Grid boxes whose
centers are part ofthe \code{SpatialPolygons} are given value 1 and all other
 grid boxes are given value 0. Land values are set to NA.
}
