% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{currSecEx}
\alias{currSecEx}
\title{Coordinates of a line segment with self-intersections}
\format{n x 2 matrix of coordiantes}
\usage{
currSecEx
}
\description{
Example of a line segment with self-intersections. We will use it
to demonstrate the \code{untwistSec} function.
}
\examples{
data(currSecEx)
head(currSecEx)
}
\keyword{datasets}
