% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitContours.R
\name{dist_mat}
\alias{dist_mat}
\title{Compute 'distances' among $n$ lines}
\usage{
dist_mat(n_lines)
}
\arguments{
\item{n_lines}{number of lines in matrix}
}
\value{
Matrix of `distances' among the indices
}
\description{
Creates a matrix specifying how difference indices are among an
ordered set of indices. For lines with indices \eqn{i} and \eqn{j},
the 'distance' computed is \eqn{|i -j|}. Results are used to define covariance.
}
