% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{ecmwf_bin}
\alias{ecmwf_bin}
\title{Binary predictions from ECMWF ensemble, September 1993-2018}
\format{array of dimension of 26 x 304 x 448 (corresponding to year x
        longitude x 448 latitude)}
\usage{
ecmwf_bin
}
\description{
The object \code{ensemble_bin} is a binary array indicating if at least half of
the ensemble members have sea ice concentrations of at least 15\% from
September 1993-2018. The predictions are from the European Center
for Medium-Range Weather Forecasts (ECMWF) ensemble at a 2.5-month
lead time. They have been converted to a Polar stereographic grid.
}
\examples{
data(ecmwf_bin)
}
\references{
Copernicus Climate Change Service (2019). Description of the c3s
            seasonal multi-system. \url{https://confluence.ecmwf.int/display/COPSRV/Description+of+the+C3S+seasonal+multi-system}

            Sea Ice Prediction Network (2019). Sea ice prediction network
            predictability portal. \url{https://atmos.uw.edu/sipn/.}
}
\keyword{datasets}
