% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_coords}
\alias{extract_coords}
\title{Function to extract coordinates.}
\usage{
extract_coords(x)
}
\arguments{
\item{x}{\code{SpatialLines} or \code{SpatialPolygons} object}
}
\value{
n x 2 matrix of coordinates
}
\description{
Function to extract coordinates from a \code{SpatialLines} object.
If there are breaks in the line, this function connects the closest points to
create one line. Note: This differs from the function \code{getCoords} in
that the ordering of the points is considered.
}
\examples{
coords <- extract_coords(reg_info$regions[[3]])
par(mfrow = c(1, 2))
plot(reg_info$regions[[3]], main = "Polygon Object")
plot(coords, type = "p", main = "Coordinates", pch = 20)
}
