% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPolys.R
\name{find_trans}
\alias{find_trans}
\title{Find transition points (points at the start/end of polygons)}
\usage{
find_trans(pts, r, reg_info, close = 12.5)
}
\arguments{
\item{pts}{matrix with two columns giving coordinates of the points}

\item{r}{integer specifying the region}

\item{reg_info}{a \code{reg_info} list (see documentation for \code{reg_info})}

\item{close}{how close a point must be to the line to count
as being on it, defaults to 12.5}
}
\description{
Find transition points (points at the start/end of polygons)
}
