% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{full}
\alias{full}
\title{Sample collection of completely ice-filled regions}
\format{An object of class \code{SpatialPolygons} of length 1.}
\usage{
data(full)
}
\description{
Example of a collection of several completely ice-filled regions stored
as a single \code{SpatialPolygons} object
}
\examples{
data(full)
}
\keyword{datasets}
