% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_dist}
\alias{get_dist}
\title{Find euclidean distance}
\usage{
get_dist(p1, p2)
}
\arguments{
\item{p1}{vector giving the x and y coordinate pair for the first point}

\item{p2}{vector giving the x and y coordinate pair for the second point}
}
\value{
distance value
}
\description{
Finds the euclidean distance between two points (ignoring projection)
}
\examples{
get_dist(c(1, 2), c(3, 4))
}
