% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_ind}
\alias{get_ind}
\title{Find indices in matrix}
\usage{
get_ind(coords, xmn = -3850, ymn = -5350)
}
\arguments{
\item{coords}{coordinates of interest}

\item{xmn}{min x (defaults to value for Northern Polar stereographic grid:
-3850)}

\item{ymn}{min y (defaults to value for Northern Polar stereographic grid:
-5350)}
}
\value{
n x 2 matrix of coordinates on a 304 x 448 grid
}
\description{
Function to find to which matrix indices coordinates correspond (on a 304 x
 448 grid)
}
\examples{
dat <- matrix(nrow = 2, ncol = 2, data = c(-2000, 0, 300, 1000))
get_ind(dat)
}
