% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{int_line}
\alias{int_line}
\title{Space points along a line}
\usage{
int_line(pred_l, obs_l, plotting = FALSE)
}
\arguments{
\item{pred_l}{predicted line (n1 x 2 matrix of coordinates)}

\item{obs_l}{predicted line (n2 x 2 matrix of coordinates)}

\item{plotting}{boolean indicating whether maps should be plotted}
}
\value{
n x 2 matrix of evenly-spaced coordinates
}
\description{
The function evenly spaces the number of points that are on one line,
\code{pred_l}, on a different line, \code{obs_l}
}
\examples{
line_space <- int_line(predLEx, obsLEx, plotting = TRUE)
}
