% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPolys.R
\name{inter_start_line}
\alias{inter_start_line}
\title{Add points where line connecting point sequence crosses start_line}
\usage{
inter_start_line(r, pts, reg_info)
}
\arguments{
\item{r}{region number}

\item{pts}{matrix with two columns giving coordinates of the points}

\item{reg_info}{a \code{reg_info} list (see documentation for \code{reg_info})}
}
\description{
Add points where line connecting point sequence crosses start_line
}
