% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{interEx}
\alias{interEx}
\title{Example of a line that contains self-intersections}
\format{n x 2 matrix of coordinates}
\usage{
interEx
}
\description{
Example of a line that contains self-intersections. We will use it to
demonstrate the functions that address these intersections.
}
\examples{
data(interEx)
plot(interEx)
}
\keyword{datasets}
