% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPolys.R
\name{interp_new_pts}
\alias{interp_new_pts}
\title{Interpolate along region boundaries}
\usage{
interp_new_pts(r, new_pts, reg_info, end = TRUE, close = 12.5)
}
\arguments{
\item{r}{integer indicating for which region the contours are being generated}

\item{new_pts}{coordinates of the contour}

\item{reg_info}{a \code{reg_info} list (see documentation for \code{reg_info})}

\item{end}{indicator determining if the points are being interpolated on
the ending coordinates or the starting coordinates. Defaults to \code{TRUE}.}

\item{close}{how close a point must be to the line to count
as being on it, defaults to 12.5}
}
\description{
Interpolate contour points that are very close or on the region boundaries.
}
