% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{keep_line}
\alias{keep_line}
\title{Keep only spatial lines}
\usage{
keep_line(my_poly)
}
\arguments{
\item{my_poly}{\code{SpatialCollections}, \code{SpatialPolygons},
\code{SpatialPoints}, or \code{SpatialLines} object}
}
\value{
\code{SpatialPolygons} object
}
\description{
Keep only \code{SpatialLines} from a spatial object.
}
\examples{
par(mfrow = c(1, 2))
plot(spatialCollEx, col = "blue", main = "Spatial Collections Object")
line_only <- keep_line(spatialCollEx)
plot(line_only, col = "blue", main = "Spatial Line Only")
}
