% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPolys.R
\name{make_polygons}
\alias{make_polygons}
\title{Create polygon from mapped points}
\usage{
make_polygons(r, my_end, poly_name = "unspecified", loop_r)
}
\arguments{
\item{r}{integer specifying the region of current interest}

\item{my_end}{n x 2  list of mapped points, i.e. the points to which the polygon should extend}

\item{poly_name}{character string to name the new polygon (defaults to "unspecified")}

\item{loop_r}{boolean indicating whether the points are going in a loop}
}
\value{
\code{SpatialPolygons} object created from the mapped points
}
\description{
Create a new polygon from the coordinates of mapped points
}
\examples{
new_poly <- make_polygons(r = 5, my_end = mappedPoints, loop_r = FALSE)
plot(new_poly)
}
