% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{map_curr_1}
\alias{map_curr_1}
\title{Sample output from \code{get_map}}
\format{An object of class \code{matrix} with 130 rows and 6 columns.}
\usage{
data(map_curr_1)
}
\description{
Example output from the \code{get_map} function for the Central Arctic region.
}
\examples{
data(map_curr_1)
}
\keyword{datasets}
