% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{mappedPoints}
\alias{mappedPoints}
\title{Example of mapped points}
\format{matrix of 1027 x 2}
\usage{
mappedPoints
}
\description{
Example of a set of mapped points organized as an n x 2 matrix of coordinates.
This is used to demonstrate the \code{makePolygons} function.
}
\examples{
data(mappedPoints)
head(mappedPoints)
plot(mappedPoints, type = "l")
}
\keyword{datasets}
