% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genContours.R
\name{merge_conts}
\alias{merge_conts}
\title{Merge contours}
\usage{
merge_conts(conts, full)
}
\arguments{
\item{conts}{list of contours organized as a list of regions by a list of
years by a list of samples}

\item{full}{\code{SpatialPolygons} object for area to be included in all
generated contours}
}
\value{
Returns a list of contours organized as a list of years by a list of
 samples
}
\description{
Merge generated contours for all regions together
}
