% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{merged}
\alias{merged}
\title{Sample list of contours}
\format{An object of class \code{list} of length 10.}
\usage{
data(merged)
}
\description{
Example list of ten contours in the form of \code{SpatialPolygons} objects
}
\examples{
data(merged)
}
\keyword{datasets}
