% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{obs_9_2008}
\alias{obs_9_2008}
\title{Observed sea ice for September 2008}
\format{array}
\usage{
data(obs_9_2008)
}
\description{
Array of dimension longitude by latitude. Binary indicate of whether
sea ice concentration of at least 15\% was observed. Computed from NASA
Boostrap sea ice concentration product (Comiso 2017).
}
\references{
Comiso, J., 2017: Bootstrap sea ice concentrations
            from Nimbus-7 SMMR and DMSP SSM/I-SSMIS. version 3.
            Boulder, Colorado USA: NASA National Snow and Ice Data Center
            Distributed Active Archive Center.
            doi: \url{https://doi.org/10.5067/7Q8HCCWS4I0R}
}
\keyword{datasets}
