% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{obs_maps}
\alias{obs_maps}
\title{Output from \code{create_mapping} function}
\format{An object of class \code{list} of length 4.}
\usage{
data(obs_maps)
}
\description{
Sample output from the \code{create_mapping} function for observations
from September 1993-2007. It is a list of four objects. The first two items
in the list, \code{start_year} and \code{end_year}, give the first and last
year that were mapped. The second two items, \code{obs_list} and
\code{pred_list}, are lists of arrays with one 3-dimensional array for each
region. The first dimension is for the  year. The other two dimensions are
for the fixed points' y-coordinates, the mapped points' x-coordinates, the
mapped points' y-coordinates, the length of the mapping vectors in the
x-direction, the length of the vectors in the y-direction, and the angles of
the mapping vectors.
}
\examples{
data(obs_maps)
}
\keyword{datasets}
