% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{obsLEx}
\alias{obsLEx}
\title{Coordinates of an observed line segment}
\format{n x 2 matrix of coordinates}
\usage{
obsLEx
}
\description{
Example of the coordinates for an observed line segment. We will use it to
demonstrate the \code{intLine} function.
}
\examples{
data(obsLEx)
head(obsLEx)
}
\keyword{datasets}
