% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{pars_1}
\alias{pars_1}
\title{Sample parameter information for generating a contour}
\format{An object of class \code{list} of length 2.}
\usage{
data(pars_1)
}
\description{
Example list with two elements, \code{mu_est} and \code{sigma_est}, which
give the mean and covariance from which an example contour can be generated
}
\examples{
data(pars_1)
}
\keyword{datasets}
