% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{ppe_9_2008}
\alias{ppe_9_2008}
\title{Post-procesed ensemble forecast for 2008}
\format{array}
\usage{
data(ppe_9_2008)
}
\description{
Array of dimension year by longitude by latitude.The initial forecasts
are from the European Center for Medium-Range Weather Forecasts (ECMWF)
ensemble at a 2.5-month lead time. They have been converted to a Polar
stereographic grid. Model output is available from
the Sea Ice Prediction Network Predicatability Portal or the Copernicus
Climate Change Service data store.
}
\references{
Copernicus Climate Change Service (2019). Description of the c3s
            seasonal multi-system. \url{https://confluence.ecmwf.int/display/COPSRV/Description+of+the+C3S+seasonal+multi-system}

            Sea Ice Prediction Network (2019). Sea ice prediction network
            predictability portal. \url{https://atmos.uw.edu/sipn/.}
}
\keyword{datasets}
