% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{pred_maps}
\alias{pred_maps}
\title{Computed mappings for predictions for September 1993-2008}
\format{An object of class \code{list} of length 4.}
\usage{
data(pred_maps)
}
\description{
Output of the \code{create_mapping} function for September 1993-2008 using
predictions from the European Center for Medium-Range Weather Forecasts
(ECMWF) ensemble converted to a Polar stereographic grid.
}
\examples{
data(pred_maps)
}
\references{
Copernicus Climate Change Service (2019). Description of the c3s seasonal multi-system.\url{https://confluence.ecmwf.int/display/COPSRV/Description+of+the+C3S+seasonal+multi-system}

            Sea Ice Prediction Network (2019). Sea ice prediction network
            predictability portal. \url{https://atmos.uw.edu/sipn/.}
}
\keyword{datasets}
