% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{predLEx}
\alias{predLEx}
\title{Coordinates of a predicted line segment}
\format{n x 2 matrix of coordinates}
\usage{
predLEx
}
\description{
Example of the coordinates for a predicted line segment. We will use it to
demonstrate the \code{intLine} function.
}
\examples{
data(predLEx)
head(predLEx)
}
\keyword{datasets}
