% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{prop_area}
\alias{prop_area}
\title{Proportion of total area by grid box}
\format{array}
\usage{
data(prop_area)
}
\description{
Matrix of dimension longitude by latitude. Elements give the proportion of
the total area (within the seas of the Arctic) in each grid box. The sum
of all elements is 1.
}
\keyword{datasets}
