% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPolys.R
\name{pt_line_inter}
\alias{pt_line_inter}
\title{Check if a point crosses a line segment}
\usage{
pt_line_inter(pt_to_test, fixed_pts)
}
\arguments{
\item{pt_to_test}{numeric vector of length two giving the point to test}

\item{fixed_pts}{matrix of dimension 2 by 2 giving the line segment to test}
}
\description{
Check if a point crosses a line segment
}
