% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPolys.R
\name{sec_to_interp}
\alias{sec_to_interp}
\title{Interpolate a section of line}
\usage{
sec_to_interp(p1 = NULL, p2 = NULL, bd_r, loop_r = FALSE)
}
\arguments{
\item{p1}{vector of length two giving the coordinates of the first point}

\item{p2}{vector length two giving the coordinates of the second point}

\item{bd_r}{matrix with two columns giving the fixed line on which to interpolate}

\item{loop_r}{boolean indicating whether the points are going in a loop}
}
\description{
Interpolate a section of line
}
\details{
If only p1 is given the point is assumed to be the first in the
sequence. If only p2 is given the point is assumed to be the last point
in the sequence
}
