% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{sipSep2008}
\alias{sipSep2008}
\title{Ensemble estimated sea ice probability September 2008 (lead time 2.5 months)}
\format{matrix of dimension 304 x 448 (longitude x latitude)}
\usage{
sipSep2008

sipSep2008
}
\description{
The object \code{sip2006_2007} is an array of the sea ice probability
predicted  from the European Center for Medium-Range Weather Forecasts
(ECMWF) ensemble converted to a Polar stereographic grid.

The object \code{sipSep2008} is an array of the proportion of ensemble
members that have sea ice concentrations of at least 15\%. The predictions are
from the European Centerfor Medium-Range Weather Forecasts (ECMWF) ensemble
converted to the Polar stereographic grid.
}
\references{
Copernicus Climate Change Service (2019). Description of the c3s
            seasonal multi-system. \url{https://confluence.ecmwf.int/display/COPSRV/Description+of+the+C3S+seasonal+multi-system}

            Sea Ice Prediction Network (2019). Sea ice prediction network
            predictability portal. \url{https://atmos.uw.edu/sipn/.}#' @examples
data(sipSep2008)
dim(sipSep2008)

Copernicus Climate Change Service (2019). Description of the c3s
            seasonal multi-system.\url{https://confluence.ecmwf.int/display/COPSRV/Description+of+the+C3S+seasonal+multi-system}

            Sea Ice Prediction Network (2019). Sea ice prediction network
            predictability portal. \url{https://atmos.uw.edu/sipn/.}
}
\keyword{data(sipSep2008)}
\keyword{datasets}
\keyword{dim(sipSep2008)}
