% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\docType{data}
\name{spatialCollEx}
\alias{spatialCollEx}
\title{Spatial collection example}
\format{\code{SpatialCollections} object}
\usage{
spatialCollEx
}
\description{
Example of a \code{SpatialCollections} object that contains a
\code{SpatialPolygons} object and a \code{SpatialLines} object
}
\examples{
data(SpatialCollEx)
plot(spatialCollEx)
plot(spatialCollEx@lineobj, col = "red", add = TRUE)
plot(spatialCollEx@polyobj, col = "blue", add = TRUE)
}
\keyword{datasets}
