% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitContours.R
\name{to_fit}
\alias{to_fit}
\title{Identify fully ice-covered and ice-free regions}
\usage{
to_fit(y_obs, reg_info)
}
\arguments{
\item{y_obs}{list of y values outputted from \code{y_obs} function}

\item{reg_info}{a \code{reg_info} list (see documentation for \code{reg_info})}
}
\description{
Determine which regions are completely ice-filled (full) or ice-free (empty)
in all years in the training period. Also, make the polygon corresponding to
regions that are fully ice-covered.
}
