% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitContours.R
\name{train_ind}
\alias{train_ind}
\title{Find indices on which to train contour model}
\usage{
train_ind(maps)
}
\arguments{
\item{maps}{output of a \code{create_mapping} object}
}
\description{
Identify the years on which to train accounting for years with missing data
}
