% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genContours.R
\name{ts_adj_mu}
\alias{ts_adj_mu}
\title{Trend Adjustment For \code{mu}}
\usage{
ts_adj_mu(obs_list, forecast_year, train_start_year, train_end_year)
}
\arguments{
\item{obs_list}{partial output of \code{get_map function},
\code{maps$obs_list[[r]]}, where r is the region of interest}

\item{forecast_year}{year to be forecast}

\item{train_start_year}{first year in training period}

\item{train_end_year}{last year in training period}
}
\value{
vector of the length of the number of lines in the mapping that
represent by what factor each estimated \code{mu} should be adjusted
}
\description{
Trend Adjustment For \code{mu}
}
