% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF.R
\name{wght_mod}
\alias{wght_mod}
\title{Function to weight two models}
\usage{
wght_mod(w, mod1, mod2)
}
\arguments{
\item{w}{weight on model 1}

\item{mod1}{array with estimated sea ice probability from model 1. Dimensions
are nuumber of training years x lon x lat.}

\item{mod2}{array with estimated sea ice probability from model 1. Dimensions
are nuumber of training years x lon x lat.}
}
\description{
Function to weight two models
}
\examples{
\dontrun{
weight <- fit_weights(mod1 = clim_9_2005_2007, mod2 = ppe_9_2005_2007,
obs = obs_9_2005_2007, prop_area = prop_area)
wght_mod(w = weight, mod1 = clim_9_2008, mod2 = ppe_9_2008)
}
}
