% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitContours.R
\name{y_obs}
\alias{y_obs}
\title{Compute \eqn{y}}
\usage{
y_obs(maps, reg_info)
}
\arguments{
\item{maps}{output of the \code{create_mapping} function}

\item{reg_info}{a \code{reg_info} list (see documentation for \code{reg_info})}
}
\value{
List of matrices, one per region, giving the observed \eqn{y} values.
        Each row corresponds to the lines in \eqn{L} and each column corresponds
        to  a training year
}
\description{
Compute y values from the output of the \code{create_mapping} object
}
\examples{
y_obs <- y_obs(maps = obs_maps, reg_info)
}
